third_party/WebKit/LayoutTests/http/tests/websocket/error-event-ready-state-expected.txt

CONSOLE ERROR: line 22: WebSocket connection to 'ws://127.0.0.1:8880/non-existent-url' failed: Error during WebSocket handshake: Unexpected response code: 404
Test that readyState is CLOSED within onerror event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test start (Port 8880)
PASS ws.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
